// database.types.ts
export type Database = {
	public: {
		Tables: {
			comments: {
				Row: {
					content: string | null;
					created_at: string;
					creatorid: string | null;
					id: string;
					likes: string[] | null;
					postid: string | null;
				};
				Insert: {
					content?: string | null;
					created_at?: string;
					creatorid?: string | null;
					id?: string;
					likes?: string[] | null;
					postid?: string | null;
				};
				Update: {
					content?: string | null;
					created_at?: string;
					creatorid?: string | null;
					id?: string;
					likes?: string[] | null;
					postid?: string | null;
				};
				Relationships: [
					{
						foreignKeyName: "comments_creatorid_fkey";
						columns: ["creatorid"];
						isOneToOne: false;
						referencedRelation: "profiles";
						referencedColumns: ["id"];
					},
					{
						foreignKeyName: "comments_postid_fkey";
						columns: ["postid"];
						isOneToOne: false;
						referencedRelation: "posts";
						referencedColumns: ["id"];
					},
				];
			};
			posts: {
				Row: {
					banner_url: string | null;
					content: string | null;
					created_at: string;
					creator_id: string | null;
					id: string;
					label: string;
					label_color: string;
					likes: string[] | null;
					title: string;
				};
				Insert: {
					banner_url?: string | null;
					content?: string | null;
					created_at?: string;
					creator_id?: string | null;
					id?: string;
					label?: string;
					label_color?: string;
					likes?: string[] | null;
					title: string;
				};
				Update: {
					banner_url?: string | null;
					content?: string | null;
					created_at?: string;
					creator_id?: string | null;
					id?: string;
					label?: string;
					label_color?: string;
					likes?: string[] | null;
					title?: string;
				};
				Relationships: [
					{
						foreignKeyName: "posts_creator_id_fkey";
						columns: ["creator_id"];
						isOneToOne: false;
						referencedRelation: "profiles";
						referencedColumns: ["id"];
					},
				];
			};
			profiles: {
				Row: {
					avatar_url: string;
					bio: string | null;
					dept: string | null;
					email: string | null;
					full_name: string | null;
					id: string;
					title: string | null;
					updated_at: string;
					username: string | null;
				};
				Insert: {
					avatar_url?: string;
					bio?: string | null;
					dept?: string | null;
					email?: string | null;
					full_name?: string | null;
					id: string;
					title?: string | null;
					updated_at?: string;
					username?: string | null;
				};
				Update: {
					avatar_url?: string;
					bio?: string | null;
					dept?: string | null;
					email?: string | null;
					full_name?: string | null;
					id?: string;
					title?: string | null;
					updated_at?: string;
					username?: string | null;
				};
				Relationships: [];
			};
		};
		Views: {
			[_ in never]: never;
		};
		Functions: {
			[_ in never]: never;
		};
		Enums: {
			[_ in never]: never;
		};
		CompositeTypes: {
			[_ in never]: never;
		};
	};
};

type PublicSchema = Database[Extract<keyof Database, "public">];

export type Tables<
	PublicTableNameOrOptions extends
		| keyof (PublicSchema["Tables"] & PublicSchema["Views"])
		| { schema: keyof Database },
	TableName extends PublicTableNameOrOptions extends { schema: keyof Database }
		? keyof (Database[PublicTableNameOrOptions["schema"]]["Tables"] &
				Database[PublicTableNameOrOptions["schema"]]["Views"])
		: never = never,
> = PublicTableNameOrOptions extends { schema: keyof Database }
	? (Database[PublicTableNameOrOptions["schema"]]["Tables"] &
			Database[PublicTableNameOrOptions["schema"]]["Views"])[TableName] extends {
			Row: infer R;
		}
		? R
		: never
	: PublicTableNameOrOptions extends keyof (PublicSchema["Tables"] &
				PublicSchema["Views"])
		? (PublicSchema["Tables"] &
				PublicSchema["Views"])[PublicTableNameOrOptions] extends {
				Row: infer R;
			}
			? R
			: never
		: never;
// ...
export type Enums<
	PublicEnumNameOrOptions extends
		| keyof PublicSchema["Enums"]
		| { schema: keyof Database },
	EnumName extends PublicEnumNameOrOptions extends { schema: keyof Database }
		? keyof Database[PublicEnumNameOrOptions["schema"]]["Enums"]
		: never = never,
> = PublicEnumNameOrOptions extends { schema: keyof Database }
	? Database[PublicEnumNameOrOptions["schema"]]["Enums"][EnumName]
	: PublicEnumNameOrOptions extends keyof PublicSchema["Enums"]
		? PublicSchema["Enums"][PublicEnumNameOrOptions]
		: never;

export type CompositeTypes<
	PublicCompositeTypeNameOrOptions extends
		| keyof PublicSchema["CompositeTypes"]
		| { schema: keyof Database },
	CompositeTypeName extends PublicCompositeTypeNameOrOptions extends {
		schema: keyof Database;
	}
		? keyof Database[PublicCompositeTypeNameOrOptions["schema"]]["CompositeTypes"]
		: never = never,
> = PublicCompositeTypeNameOrOptions extends { schema: keyof Database }
	? Database[PublicCompositeTypeNameOrOptions["schema"]]["CompositeTypes"][CompositeTypeName]
	: PublicCompositeTypeNameOrOptions extends keyof PublicSchema["CompositeTypes"]
		? PublicSchema["CompositeTypes"][PublicCompositeTypeNameOrOptions]
		: never;